+++ HTML/include/include_proj_list_projets.php
@@ -1,7 +1,7 @@
 <?php
 // File: include/include_proj_list_project.php
 // Function: List all projects, used in proj_list.php
-// Revision date: 2026-09-21
+// Revision date: 2026-09-23
 // Revised by: Mikael Karlsson
 // This file is distributed under the license:
 // Creative Commons Attribution-NonCommercial-ShareAlike 3.0 Unported License.
@@ -43,31 +43,48 @@ class Proj {
         $sql_exec = mysqli_Query($connection,$GetDataComponentsAll);
 
         while($showDetails = mysqli_fetch_array($sql_exec)) {
-            echo "<tr>";
+            // First row
+            echo '<tr>';
+            // First column
             echo '<td class="edit"><a href="proj_edit.php?proj_id=';
             echo $showDetails['project_id'];
-            echo '"><span class="fas fa-pencil-alt fa-lg"></span></a></td>';
-
-            echo "<td>";
+            echo '"><span class="fas fa-pencil-alt fa-lg"></span></a>';
+            echo '</td>'; //end first column
+            // Second column
+            echo '<td>';
             echo '<a href="proj_show.php?proj_id=';
             echo $showDetails['project_id'];
             echo '">';
             echo $showDetails['project_name'];
             echo '</a>';
-            echo "</td>";
-
-            echo "<td>";
-            $components = mysqli_query($connection,"SELECT projects_data_project_id FROM projects_data WHERE projects_data_project_id = ".$showDetails['project_id']."");
+            echo '</td>'; // end second column
+            // third column
+            echo '<td>';
+            $components = mysqli_query($connection,"SELECT projects_data_project_id, projects_data_quantity FROM projects_data WHERE projects_data_project_id = ".$showDetails['project_id']."");
             $number_components = mysqli_num_rows($components);
             if ($number_components == 0){
-                echo "-";
+                echo "0";
             }
             else{
                 echo $number_components;
             }
-            echo "</td>";
+            echo '</td>'; // end third column
+            // fourth column
+            echo '<td>';
+            if( $number_components > 0 ) { // if there are components in the project
+                $NoOfComponents = mysqli_fetch_assoc($components); // get the component quantity
+                if ( $NoOfComponents['projects_data_quantity'] == "" ) { // empty?
+                    echo "0"; // yes, display a zero
+                } else {
+                    echo $NoOfComponents['projects_data_quantity']; // display the unit count
+                }
+            } else {
+                echo "0"; // No components in the project
+            }
+            // end fourth column
+            echo '</td>';
 
-            echo "<td>";
+            echo '<td>'; // start fifth column
             $GetDataPrice = "SELECT SUM(total) FROM (SELECT projects_data_quantity * price AS total FROM projects_data JOIN `data` WHERE data.id = projects_data_component_id AND projects_data_project_id = ".$showDetails['project_id'].") AS project_total";
             $sql_exec_price = mysqli_Query($connection,$GetDataPrice) or die(mysql_error());
             $showPrice = mysqli_fetch_array($sql_exec_price);
@@ -79,7 +96,7 @@ class Proj {
                     $price = $sum;
                 }
                     echo format_currency($personal['language'],$personal['currency'], $price) ;
-            echo "</td></tr>";
+            echo '</td></tr>'; // end fifth column, end row
         }
     }
 }